Update README.md
Bus Booking Website designed to facilitate online booking and reservation of bus tickets. The platform
enables users to search for available buses, book seats, and manage their reservations. It is built using
Spring MVC for the web application framework, Hibernate for ORM (Object-Relational Mapping), and
MySQL as the database. The frontend is developed using HTML and Bootstrap for a responsive and user friendly interface.